+++ sdk/examples/client/output/build_nft_output.rs
@@ -42,16 +42,14 @@ async fn main() -> Result<()> {
     "name":"My NFT #0001"
     }"#,
     )?
-    .to_string()
-    .as_bytes()
-    .to_vec();
+    .to_string();
 
     // NftId needs to be null the first time
     let nft_output = NftOutputBuilder::new_with_minimum_storage_deposit(rent_structure, NftId::null())
         .add_unlock_condition(AddressUnlockCondition::new(address))
         .add_feature(SenderFeature::new(address))
-        .add_feature(MetadataFeature::new("mutable metadata".as_bytes().to_vec())?)
-        .add_feature(TagFeature::new("my tag".as_bytes().to_vec())?)
+        .add_feature(MetadataFeature::new("mutable metadata")?)
+        .add_feature(TagFeature::new("my tag")?)
         .add_immutable_feature(IssuerFeature::new(address))
         .add_immutable_feature(MetadataFeature::new(tip_27_immutable_metadata)?)
         .finish_output(token_supply)?;