+++ sdk/src/wallet/account/operations/participation/voting_power.rs
@@ -72,7 +72,7 @@ impl Account {
                             .address
                             .inner,
                     ))
-                    .add_feature(TagFeature::new(PARTICIPATION_TAG.as_bytes().to_vec())?)
+                    .add_feature(TagFeature::new(PARTICIPATION_TAG)?)
                     .finish_output(token_supply)?,
                 None,
             ),