+++ sdk/CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AccountInner::{incoming_transactions, transactions, pending_transactions}` don't return a `Result` anymore;
 - `AccountInner::incoming_transactions` returns a `Vec` instead of a `HashMap`;
 - `Address::try_from_bech32_with_hrp` refactored to `try_from_bech32`;
+- `{MetadataFeature, TagFeature}::new` take an `impl Into<Vec<u8>>` param;
 
 ### Removed